requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx>=0.27.0
anthropic>=0.39.0
scipy>=1.12.0
numpy>=1.26.0
apscheduler>=3.10.0
xgboost>=2.0.0
scikit-learn>=1.4.0
pandas>=2.1.0
joblib>=1.3.0
faiss-cpu>=1.7.0
rank-bm25>=0.2.2
